Chore: 불필요한 소켓 연결을 피하기 위해 판매자 채팅, 편지 목록 리스트 분리하고 내부에서 fetch #150
