+++ package.json
@@ -15,6 +15,7 @@
     "@types/stompjs": "^2.3.9",
     "axios": "^1.6.2",
     "cra-template-typescript": "1.2.0",
+    "dotenv": "^16.4.5",
     "react": "^18.2.0",
     "react-beforeunload": "^2.6.0",
     "react-cookie": "^7.0.1",
